+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/HlsMediaPeriod.java
@@ -957,6 +957,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
         .setFrameRate(variantFormat.frameRate)
         .setSelectionFlags(variantFormat.selectionFlags)
         .setRoleFlags(variantFormat.roleFlags)
+        .setColorInfo(variantFormat.colorInfo)
         .build();
   }